+++ multicalc.js
@@ -1,7 +1,3 @@
-// change the monitoring recommendation
-
-
-
 /**
  * Copyright (c) 2021
  *
@@ -12,7 +8,7 @@
  * Based on Excel Multipurpose Calculator created by Dennis Tran
  *
  * Created at     : 2021-01-15
- * Last modified  : 2021-07-18
+ * Last modified  : 2021-08-21
  */
 
 
@@ -24,10 +20,10 @@ import { displayDate, displayValue, checkValue, roundTo, getDateTime, getHoursBe
 import { default as ivig } from './modules/ivig.js';
 import { childIsObese } from './modules/growthCharts.js';
 import { getSecondDose } from './modules/seconddose.js';
+import { default as arial } from './modules/arial.js';
+// TODO: set false
 let debug = true;
 let debugDefaultTab = "initial";
-
-// CHANGED
 
 $( function(){
   $('[data-toggle="popover"]').popover({html: true});
@@ -38,11 +34,6 @@ $( function(){
   } else if ( /log/.test(location.search) ) {
     LOG.enable();
   }
-
-  // ADDED
-
-
-  // \ADDED
 
   if ( debug ){
     $("#ptage").val(60);
@@ -106,7 +97,6 @@ $( function(){
       })
     }
   }
-  // ADDED
   $('#ptage').get(0).focus();
 });
 
@@ -147,6 +137,7 @@ $(".input-auc").on('keyup', () => {
   calculate.vancoAUC();
   calculate.vancoRevision();
 });
+
 $(".input-auc-interval").on('keyup', () => {
   calculate.vancoAUC(false);
 });
@@ -154,10 +145,10 @@ $(".input-auc-interval").on('keyup', () => {
 $("#auc-reset").on('click', () => {
   calculate.vancoAUC();
 });
+
 $('#btnShowInitialPk').on('click', () => {
   $('#row--initialPkAlert').hide(50, "linear");
   $('#row--initialPkCalc').addClass('show');
-
 });
 
 $(".input-revision").on('keyup', () => {
@@ -165,33 +156,42 @@ $(".input-revision").on('keyup', () => {
   calculate.vancoRevision();
   calculate.vancoAUC();
 });
+
 $(".input-aucDates").on('keyup', () => {
   calculate.vancoAUCDates();
 });
+
 $(".input-twolevel").on('keyup', () => {
   calculate.vancoTwolevel();
 });
+
 $(".input-twolevel-interval").on('keyup', () => {
   calculate.vancoTwolevel(false);
 });
+
 $("#twolevel-reset").on('click', () => {
   calculate.vancoTwolevel();
 });
+
 $("#schwartz-k-infant").on('change', () => {
   calculate.patientData();
   calculate.vancoInitial();
   calculate.vancoRevision();
   calculate.vancoAUC();
 });
+
 $(".input-steadystate").on('keyup', () => {
   calculate.vancoSteadyStateCheck();
 });
+
 $("#seconddose-time1").on("keyup", () => {
   calculate.secondDose();
 });
+
 $("[name='seconddose-freq']").on("change", () => {
   calculate.secondDose();
 });
+
 $("#revision-goalTrough").on("change", () => {
   calculate.vancoRevision();
 })
@@ -239,7 +239,6 @@ $("#aucDates-sameInterval").on("change", (e) => {
     $("#aucDates-row--dose2").show();
     par.insertBefore(tr, pk);
     par.insertBefore(pk, spacer);
-
   }
   calculate.vancoAUCDates();
 });
@@ -663,10 +662,6 @@ const vanco = {
         txtExceeds = `<br>***Upper range of protocol exceeds ${maxDaily/1000} g/day***`;
       }
     }
-    // CHANGED
-
-
-
     if ( consider > 0 && lowDose !== highDose ) {
       considerText = `<br><i>Consider dosing closer to ${lowDose} mg.</i>`;
     }
@@ -689,9 +684,6 @@ const vanco = {
         goalTroughIndex: 0,
         monitoring: 'Draw levels for AUC calculation when at steady state<br><span class="semibold">(if therapy anticipated to be &gt;&nbsp;72&nbsp;hours)</span>'
       },
-
-
-      // CHANGED (unless kinetic outlier, monitoring rec's)
 
       pedsAuc: {
         param: 'trough',
@@ -798,16 +790,8 @@ const vanco = {
     const highSCr = scr >= 1.2;
     const highBMI = bmi > 30;
     const lowSCr = scr < 0.5;
-    // CHANGED
 
     const earlyTroughReason = `${lowCrCl ? 'CrCl &lt; 50' : ''}${lowCrCl && highSCr ? ' and ' : ''}${highSCr ? 'SCr &ge; 1.2' : ''}`;
-
-
-    // CHANGED
-
-
-
-
 
     if ( lowCrCl || highSCr ) {
       res.monitoring = `Consider pre-steady state level<br><i>(to spot check for clearance, for ${earlyTroughReason})</i>`;
@@ -960,6 +944,11 @@ const vanco = {
       vd: vd,
       ke: ke,
       auc24: auc24,
+      aucInf: aucInf,
+      aucElim: aucElim,
+      truePeak: truePeak,
+      trueTrough: trueTrough,
+      tInf: tInf,
       therapeutic: this.aucTherapeutic(auc24),
       oldDose: dose,
       oldInterval: interval,
@@ -1242,6 +1231,34 @@ const vanco = {
      let cboIndication = $("#vancoIndication")[0];
      pt.hd = cboHD.selectedIndex;
      pt.vancoIndication = cboIndication.selectedIndex;
+     let tapePt = [[
+      ["Age", displayValue('', pt.age, 0.01, " years")],
+      ["Sex", pt.sex === 0 ? "" : pt.sex],
+      ["Weight", displayValue('', pt.wt, 0.0001, " kg")],
+      ["Height", displayValue('', pt.ht, 0.0001, " cm")],
+      ["SCr", displayValue('', pt.scr, 0.0001, " mg/dL")]
+    ]];
+    if ( pt.ageContext === "adult" ){
+      tapePt.push([
+        ["Ideal body weight", displayValue('', pt.ibw, 0.1, ' kg')],
+        ["Over/Under IBW", displayValue('', pt.overUnder, 0.1, "%", '', true)],
+        ["Adjusted body weight", displayValue('', pt.adjBW, 0.1, " kg")],
+        ["Lean body weight", displayValue('', pt.lbw, 0.1, " kg")],
+        ["Body mass index", displayValue('', pt.bmi, 0.1, " kg/m²")]
+      ]);
+      tapePt.push([
+        ["CrCl (C-G Actual)", displayValue('', pt.cgActual, 0.1, " mL/min")],
+        ["CrCl (C-G Ideal)", displayValue('', pt.cgIdeal, 0.1, " mL/min")],
+        ["CrCl (C-G Adjusted)", displayValue('', pt.cgAdjusted, 0.1, " mL/min")],
+        ["CrCl (Protocol)", displayValue('', pt.crcl, 0.1, " mL/min")]
+      ]);
+    } else {
+      tapePt.push([
+        ["Body mass index", displayValue('', pt.bmi, 0.1, " kg/m²")],
+        ["CrCl (Schwartz)", displayValue('', pt.schwartz, 0.1, " mL/min")]
+      ]);
+    }
+    $("#tape--pt").html(LOG.outputTape(tapePt, "Patient Info"));
    },
    vancoInitial(){
      $("#vancoInitialLoad").html(vanco.loadingDose());
@@ -1316,6 +1333,13 @@ const vanco = {
        highlightColumns: arrViable
      });
      $("#vancoInitialPK-table").html(tableHtml);
+     let tableText = [];
+// WORKING ON THIS
+
+     // for ( let i = 0; i < arrDose.length; i++ ){
+     //   tableText.push([`${aucNew.dose[i]} mg q${newInterval}h`,
+     //        `predicted AUC = ${roundTo(aucNew.auc[i], 0.1)}, est. trough = ${roundTo(aucNew.trough[i],0.1)} mcg/mL`])
+     // }
 
    },
    createVancoTable({rows, highlightColumns} = {}){
@@ -1422,7 +1446,6 @@ const vanco = {
        $("#steadystate-timeDiff").html("");
      }
    },
-   // ADDED
    vancoAUCDates(){
      const sameInterval = $('#aucDates-sameInterval').is(':checked');
      const dose1 = getDateTime($('#aucDates-doseDate-1').val(), $('#aucDates-doseTime-1').val());
@@ -1441,8 +1464,8 @@ const vanco = {
        dose: pt.curDose,
        interval: pt.curFreq,
        peak: checkValue(+$("#auc-curPeak").val(), vanco.config.check.levelMin, vanco.config.check.levelMax),
-       peakTime: checkValue(+$("#vancoAUCPeakTime").val(), 0, 36),
-       troughTime: checkValue(+$("#vancoAUCTroughTime").val(), 0, 36),
+       peakTime: checkValue(+$("#vancoAUCPeakTime").val(), vanco.config.check.timeMin, vanco.config.check.timeMax),
+       troughTime: checkValue(+$("#vancoAUCTroughTime").val(), vanco.config.check.timeMin, vanco.config.check.timeMax),
        trough: pt.curTrough
      };
      const aucCurrent = vanco.calculateAUC(params);
@@ -1472,6 +1495,52 @@ const vanco = {
      highlightColumns: aucNew.therapeutic
    });
    $("#aucTable").html(tableHtml);
+   let tableText = [["Predicted AUC (est. trough) for New Dose", []]];
+   for ( let i = 0; i < aucNew.dose.length; i++ ){
+     tableText[0][1].push([`${aucNew.dose[i]} mg q${newInterval}h`,
+          `${arial.padArray(["9999.99", roundTo(aucNew.auc[i], 0.1)], 0)[1]} (${roundTo(aucNew.trough[i],0.1)} mcg/mL)`])
+   }
+
+   tape.auc = [
+     [
+       ["Current dose", `${params.dose} mg q${params.interval}h`],
+       ["Peak level", displayValue('', params.peak, 0.1, " mcg/mL")],
+       ["Trough level", displayValue('', params.trough, 0.1, " mcg/mL")],
+       ["Dose before trough", `${displayDate(dose1DT)}`],
+       ["Trough drawn at", `${displayDate(peakDT)}`],
+       ["Dose before peak", `${displayDate(dose2DT)}`],
+       ["Peak drawn at", `${displayDate(peakDT)}`],
+     ],
+     [
+       ["Peak time", displayValue('', params.peakTime, 0.01, " hrs")],
+       ["Trough time", displayValue('', params.troughTime, 0.01, " hrs")],
+       ["Vd", displayValue('', aucCurrent.vd, 0.01, " L")],
+       ["Infusion time", displayValue('', aucCurrent.tInf, 1, ' min')],
+       ["ke", displayValue('', aucCurrent.ke, 0.0001, ` hr^-1`)],
+       ["True peak", displayValue('', aucCurrent.truePeak, 0.1, ' mcg/mL')],
+       ["True trough", displayValue('', aucCurrent.trueTrough, 0.1, ' mcg/mL')],
+       ["AUC (inf)", displayValue('', aucCurrent.aucInf, 0.1)],
+       ["AUC (elim)", displayValue('', aucCurrent.aucElim, 0.1)],
+     ],
+     [
+       ["AUC₂₄", `${displayValue('', aucCurrent.auc24, 0.1)} (${aucCurrent.therapeutic})`],
+       [
+         `Trough goal`,
+         `${roundTo(aucCurrent.goalTroughLow, 0.1)} &ndash; ${roundTo(aucCurrent.goalTroughHigh, 0.1)} mcg/mL`],
+       ],
+     ]
+     /********************************************
+         #     #
+          #   #   ####  #    #      ##   #####  ######    #    # ###### #####  ######
+           # #   #    # #    #     #  #  #    # #         #    # #      #    # #
+            #    #    # #    #    #    # #    # #####     ###### #####  #    # #####
+            #    #    # #    #    ###### #####  #         #    # #      #####  #
+            #    #    # #    #    #    # #   #  #         #    # #      #   #  #
+            #     ####   ####     #    # #    # ######    #    # ###### #    # ######
+      ********************************************/
+     $("#tape--auc").html(LOG.outputTape(tape.auc, "AUC Dosing"));
+     $("#tape--auc").append(LOG.outputTape(tableText));
+   },
  },
  vancoTwolevel(resetInterval = true){
    const { levelMin, levelMax, freqMin, freqMax, doseMin, doseMax } = vanco.config.check;