+++ changelog.md
@@ -8,10 +8,13 @@
 - Input date and time of levels and doses instead of just time in hours
 - Color change of Bayesian calculator alert when BMI > 30
 - Output of inputs, outputs, and calculation steps
+- Hour calculator for AUC dosing (modal)
 
 ### Changed
 - Added year to displayDate Util function
 - VMFH Logo as svg instead of png
+- Focus on age input when page loads
+- Modified monitoring recommendations to clarify when to draw trough